requirements.txt
asgiref==3.8.1
dj-database-url==1.3.0
Django==4.2.25
django-cors-headers==4.4.0
djangorestframework==3.15.2
environs==11.2.1
marshmallow==3.22.0
packaging==25.0
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.13.2
django-filter==23.2
requests==2.32.3
django-allauth[socialaccount]==65.13.1